refactor: split bootstrap dashboard, pipeline, and charts into modules
Break the monolithic components into per-feature directories:
- bootstrap-dashboard/ (per-tab components + data hook + metrics)
- bootstrap-pipeline/ (stage rail/canvas/detail, stages, reduced-motion hook)
- chart/ (chart-core + per-chart-type components)
and update charts.tsx, bootstrap-page.tsx, bootstrap-pipeline.tsx accordingly.
