+++ frontend/src/components/bootstrap-dashboard/overview-tab.tsx
@@ -1,0 +1,103 @@
+import { Boxes, Clock, Database, FileCode2, Network, Share2 } from "lucide-react";
+
+import { StatCards } from "@/components/bootstrap/stat-cards";
+import { RepoScatterChart, type ScatterPoint } from "@/components/charts";
+import { SectionShell } from "@/components/section-shell";
+import { Card, CardContent } from "@/components/ui/card";
+import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
+import type { BootstrapSummaryDoc } from "@/data/bootstrap-types";
+import { formatCount, formatDuration, formatNumber, formatPercent, modelDisplayName } from "@/lib/format";
+
+import { coveragePct, perKiloLine } from "./metrics";
+
+export function OverviewTab({
+  aggregate,
+  sizeAxis,
+  sizeScatter,
+  onSizeAxisChange
+}: {
+  aggregate: BootstrapSummaryDoc["aggregate"];
+  sizeAxis: "files" | "lines";
+  sizeScatter: ScatterPoint[];
+  onSizeAxisChange: (value: "files" | "lines") => void;
+}) {
+  return (
+    <>
+      <StatCards
+        stats={[
+          {
+            label: "Repositories",
+            value: formatCount(aggregate.totals.repos),
+            hint: `${formatCount(aggregate.totals.succeeded)} bootstrapped ok`,
+            explanation: "Number of pinned repositories included in this bootstrap benchmark run.",
+            icon: Database
+          },
+          {
+            label: "Chunks / 1k LOC",
+            value: formatNumber(perKiloLine(aggregate.totals.chunks, aggregate.totals.indexed_lines)),
+            hint: "per 1,000 indexed lines",
+            explanation:
+              "Chunks generated by Cortex per 1,000 indexed lines of code. LOC means lines of code; this uses indexed lines, not total repository lines.",
+            icon: Boxes
+          },
+          {
+            label: "Edges / 1k LOC",
+            value: formatNumber(perKiloLine(aggregate.totals.edges, aggregate.totals.indexed_lines)),
+            hint: "per 1,000 indexed lines",
+            explanation:
+              "Graph relation edges loaded per 1,000 indexed lines of code. LOC means lines of code; this normalizes graph density by the code Cortex processed.",
+            icon: Network
+          },
+          {
+            label: "Cortex coverage",
+            value: formatPercent(coveragePct(aggregate.totals.indexed_lines, aggregate.totals.tracked_lines)),
+            hint: "indexed lines / repo lines",
+            explanation:
+              "Share of tracked repository lines that Cortex indexed. Lower values usually mean generated, vendored, binary, or unsupported files were skipped.",
+            icon: FileCode2
+          },
+          {
+            label: "Embedding models",
+            value: String(aggregate.totals.models.length),
+            hint: aggregate.totals.models.map(modelDisplayName).join(", "),
+            explanation: "Number of embedding models measured in this run.",
+            icon: Share2
+          },
+          {
+            label: "Time / 1k LOC",
+            value: formatDuration(perKiloLine(aggregate.totals.duration_ms, aggregate.totals.indexed_lines)),
+            hint: "bootstrap wall-clock",
+            explanation:
+              "Total bootstrap wall-clock time per 1,000 indexed lines of code. LOC means lines of code, and the denominator is Cortex indexed lines.",
+            icon: Clock
+          }
+        ]}
+      />
+      <SectionShell
+        title="Repo size vs chunking output"
+        description="Relationship between repository size and the number of chunks the ingest phase produced. Switch the size axis between all tracked files and the lines actually subject to cortex processing."
+        headerAside={
+          <ToggleGroup
+            type="single"
+            value={sizeAxis}
+            onValueChange={(value) => value && onSizeAxisChange(value as "files" | "lines")}
+          >
+            <ToggleGroupItem value="files">Tracked files</ToggleGroupItem>
+            <ToggleGroupItem value="lines">Indexed lines</ToggleGroupItem>
+          </ToggleGroup>
+        }
+      >
+        <Card>
+          <CardContent className="pt-6">
+            <RepoScatterChart
+              points={sizeScatter}
+              xLabel={sizeAxis === "files" ? "tracked files" : "indexed lines"}
+              yLabel="chunks"
+              height={340}
+            />
+          </CardContent>
+        </Card>
+      </SectionShell>
+    </>
+  );
+}
